\name{r2hpcc-package}
\alias{r2hpcc-package}
\alias{r2hpcc}
\docType{package}
\title{
\packageTitle{r2hpcc}
}
\description{
\packageDescription{r2hpcc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{r2hpcc}
\packageIndices{r2hpcc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{r2hpcc}

Maintainer: \packageMaintainer{r2hpcc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
