% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteDropZoneFile.R
\name{r2hpcc.DeleteDropZoneFile}
\alias{r2hpcc.DeleteDropZoneFile}
\title{This function uploads file to specified table in the loading zone}
\usage{
r2hpcc.DeleteDropZoneFile(conn, dropZoneFileName, loadingZonePath)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{dropZoneFileName}{- table name where file will be stored. If table exists - it will be overwrtitten}

\item{loadingZonePath}{- loading zone path}
}
\value{
- returns result of file uploading
}
\description{
This function uploads file to specified table in the loading zone
}
