% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFUArrayAction.R
\name{r2hpcc.DeleteLogicalFiles}
\alias{r2hpcc.DeleteLogicalFiles}
\title{Title}
\usage{
r2hpcc.DeleteLogicalFiles(conn, logicalFiles)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{workunits}{- list of workunits to delete}
}
\value{
- status of processed operation
}
\description{
Title
}
