% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WUAction.R
\name{r2hpcc.DeleteWsWorkunits}
\alias{r2hpcc.DeleteWsWorkunits}
\title{Title}
\usage{
r2hpcc.DeleteWsWorkunits(conn, workunits)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{workunits}{- list of workunits to delete}
}
\value{
- status of processed operation
}
\description{
Title
}
