% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echo.R
\name{r2hpcc.Echo}
\alias{r2hpcc.Echo}
\title{Calls the executesql function of ws_sql}
\usage{
r2hpcc.Echo(conn, request)
}
\arguments{
\item{conn}{- hpcc connection information}

\item{request}{- echo message}
}
\description{
Takes in two parameters HPCC SQL statement and HPCC connection object
Returns a data frame.
If the function eencountered any error while executing the query, 2nd element will have a value of -1 and 1st element, the error message.
}
