% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exception.R
\name{r2hpcc.Exception}
\alias{r2hpcc.Exception}
\title{Calls the executesql function of ws_sql}
\usage{
r2hpcc.Exception(conn, SOAPResponse)
}
\arguments{
\item{con}{- hpcc connection information}

\item{query}{- SQL statement to execute}

\item{limit}{- limit number of rows to retrieve (defaults to connection level if not provided)}
}
\description{
Takes in two parameters HPCC SQL statement and HPCC connection object
Returns a data frame.
If the function eencountered any error while executing the query, 2nd element will have a value of -1 and 1st element, the error message.
}
