% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecutePreparedSQL.R
\name{r2hpcc.ExecutePreparedSQL2}
\alias{r2hpcc.ExecutePreparedSQL2}
\title{This method executes a previously created parameterized SQL query.}
\usage{
r2hpcc.ExecutePreparedSQL2(conn, workunitId, variables = NULL,
  supressResults = 0, suppressXMLSchema = 1, timeout = -1,
  resultWindowStart = 0, resultWindowCount = 0)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{workunitId}{- The Workunit ID (WUID)}

\item{variables}{- List of pairs(name, value) to replace placeholders in prepared SQL}

\item{supressResults}{- If set to 1 or true, query results are not included in response}

\item{suppressXMLSchema}{- If set to 1 or true, the query result schema is not included in response}

\item{timeout}{- Timeout value in milliseconds. Use -1 for no timeout}

\item{resultWindowStart}{- For use with page-loading, the starting record to return}

\item{resultWindowCount}{- For use with page-loading, the number of records to include from the ResultWindowStart}
}
\value{
Workunit details
}
\description{
This method executes a previously created parameterized SQL query.
}
