% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDBSystemInfo.R
\name{r2hpcc.GetDBSystemInfo}
\alias{r2hpcc.GetDBSystemInfo}
\title{This method allows you to get HPCC System version information.}
\usage{
r2hpcc.GetDBSystemInfo(conn, includeAll = 1)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{includeAll}{- If set to 1 or true, all available information is returned}
}
\value{
HPCC System version information
}
\description{
This method allows you to get HPCC System version information.
}
