% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWorkunit.R
\name{r2hpcc.GetDFUWorkunit}
\alias{r2hpcc.GetDFUWorkunit}
\title{This function uploads file to specified table in the loading zone}
\usage{
r2hpcc.GetDFUWorkunit(conn, workunit)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{workunit}{}
}
\value{
- returns result of file uploading
}
\description{
This function uploads file to specified table in the loading zone
}
