% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWorkunits.R
\name{r2hpcc.GetDFUWorkunits}
\alias{r2hpcc.GetDFUWorkunits}
\title{This function returns a list of DFU workunits}
\usage{
r2hpcc.GetDFUWorkunits(conn, pageSize = 50, pageStartFrom = 0)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{pageSize}{}

\item{pageStartFrom}{}
}
\value{
- returns result of file uploading
}
\description{
This function returns a list of DFU workunits
}
