% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDropZones.R
\name{r2hpcc.GetDropZones}
\alias{r2hpcc.GetDropZones}
\title{This function performs an action on a list of workunits}
\usage{
r2hpcc.GetDropZones(conn)
}
\arguments{
\item{conn}{- HPCC connection information}
}
\value{
- status of processed operation
}
\description{
This function performs an action on a list of workunits
}
