% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogicalFiles.R
\name{r2hpcc.GetLogicalFiles}
\alias{r2hpcc.GetLogicalFiles}
\title{This function returns a list of logical files}
\usage{
r2hpcc.GetLogicalFiles(conn, pageSize = 50, pageStartFrom = 0)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{pageSize}{}

\item{pageStartFrom}{}
}
\value{
- returns result of file uploading
}
\description{
This function returns a list of logical files
}
