% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRelatedIndexes.R
\name{r2hpcc.GetRelatedIndexes}
\alias{r2hpcc.GetRelatedIndexes}
\title{This function retrieves information from logical file descriptions about annotations indicating an index file which is
related to a data file.}
\usage{
r2hpcc.GetRelatedIndexes(conn, fileNames)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{fileNames}{- List of logical filenames to which the annotation is added}
}
\value{
Workunit result
}
\description{
This function retrieves information from logical file descriptions about annotations indicating an index file which is
related to a data file.
}
