% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResults.R
\name{r2hpcc.GetResults}
\alias{r2hpcc.GetResults}
\title{Calls the executesql function of ws_sql}
\usage{
r2hpcc.GetResults(conn, workunitId, suppressXMLSchema = 1,
  resultWindowStart = 0, resultWindowCount = 0)
}
\arguments{
\item{conn}{- hpcc connection information}

\item{includeAll}{- return all or limited available information}
}
\description{
Takes in two parameters HPCC SQL statement and HPCC connection object
Returns a data frame.
If the function eencountered any error while executing the query, 2nd element will have a value of -1 and 1st element, the error message.
}
