% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResults.R
\name{r2hpcc.GetResults2}
\alias{r2hpcc.GetResults2}
\title{This method allows you to get results from previously executed queries.
Use the Wuid returned from ExecuteSQL or PrepareSQL.
This method is ideal for results paging}
\usage{
r2hpcc.GetResults2(conn, workunitId, suppressXMLSchema = 1,
  resultWindowStart = 0, resultWindowCount = 0)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{workunitId}{- Workunit Id (WUID)}

\item{suppressXMLSchema}{- If set to 1 or true, the query result schema is not included in response}

\item{resultWindowStart}{- For use with page-loading, the starting record to return}

\item{resultWindowCount}{- For use with page-loading, the number of records to include from the ResultWindowStart}
}
\value{
Workunit details
}
\description{
This method allows you to get results from previously executed queries.
Use the Wuid returned from ExecuteSQL or PrepareSQL.
This method is ideal for results paging
}
