% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WUQuery.R
\name{r2hpcc.GetWsWorkunit}
\alias{r2hpcc.GetWsWorkunit}
\title{This function uploads file to specified table in the loading zone}
\usage{
r2hpcc.GetWsWorkunit(conn, workunit)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{workunit}{- workunit Id}
}
\value{
- returns result of file uploading
}
\description{
This function uploads file to specified table in the loading zone
}
