% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WUQuery.R
\name{r2hpcc.GetWsWorkunits}
\alias{r2hpcc.GetWsWorkunits}
\title{This function returns a list of WS workunits}
\usage{
r2hpcc.GetWsWorkunits(conn, pageSize = 50, pageStartFrom = 0)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{pageSize}{}

\item{pageStartFrom}{}
}
\value{
- returns result of file uploading
}
\description{
This function returns a list of WS workunits
}
