% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareSQL.R
\name{r2hpcc.PrepareSQL}
\alias{r2hpcc.PrepareSQL}
\title{Use this method to submit a free-hand SQL request for later use as a parameterized query.
This compiles the query and returns the Wuid.
This Wuid is later used to execute the query with provided input parameters using the ExecutePreparedSQL method.}
\usage{
r2hpcc.PrepareSQL(conn, sqlQuery, timeOut = -1)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{sqlQuery}{- Free-hand SQL text}

\item{timeOut}{- Timeout value in milliseconds. Use -1 for no timeout}
}
\value{
Workunit details
}
\description{
Use this method to submit a free-hand SQL request for later use as a parameterized query.
This compiles the query and returns the Wuid.
This Wuid is later used to execute the query with provided input parameters using the ExecutePreparedSQL method.
}
