% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetRelatedIndexes.R
\name{r2hpcc.SetRelatedIndexes}
\alias{r2hpcc.SetRelatedIndexes}
\title{This function adds a description to a logical file to be used as an annotation indicating an index file which is related
to a data file. This makes it available to WsSQL for use in an indexed fetch.}
\usage{
r2hpcc.SetRelatedIndexes(conn, fileName, indexes)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{fileName}{- Logical filename to which the annotation is added}

\item{indexes}{- List of indexes to add to annotation}
}
\value{
Workunit result
}
\description{
This function adds a description to a logical file to be used as an annotation indicating an index file which is related
to a data file. This makes it available to WsSQL for use in an indexed fetch.
}
