% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spray.R
\name{r2hpcc.SprayFixed}
\alias{r2hpcc.SprayFixed}
\title{This function deletes a list of logical files}
\usage{
r2hpcc.SprayFixed(conn, sourseIP, soursePath, recordSize, destinationGroup,
  namePrefix, targetName, overwrite = FALSE, compress = FALSE)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{logicalFiles}{- list of logical files}
}
\value{
- status of processed operation
}
\description{
This function deletes a list of logical files
}
