% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFetch.R
\name{r2hpcc.sqlFetch}
\alias{r2hpcc.sqlFetch}
\title{Calls the executesql function of ws_sql}
\usage{
r2hpcc.sqlFetch(Con, filename, limit)
}
\arguments{
\item{filename}{- filename to fetch from HPCC}

\item{limit}{- limit number of rows to retrieve (defaults to connection level if not provided)}

\item{con}{- hpcc connection information}
}
\description{
Takes in two parameters HPCC filename and HPCC connection object
Returns a data frame.
If the function eencountered any error while executing the query, 2nd element will have a value of -1 and 1st element, the error message.
}
