% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFetchResults.R
\name{r2hpcc.sqlFetchResults}
\alias{r2hpcc.sqlFetchResults}
\title{Calls the GetResults  function of ws_sql}
\usage{
r2hpcc.sqlFetchResults(Con, wuid, limit)
}
\arguments{
\item{wuid}{- existing HPCC WorkUnit ID}

\item{limit}{- limit number of rows to retrieve (defaults to connection level if not provided)}

\item{con}{- hpcc connection information}
}
\description{
Takes in two parameters HPCC WorkunitID and HPCC connection object
Returns a data frame.
If the function eencountered any error while executing the query, 2nd element will have a value of -1 and 1st element, the error message.
}
