% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFUWorkunitsAction.R
\name{r2hpcc.UnprotectDFUWorkunits}
\alias{r2hpcc.UnprotectDFUWorkunits}
\title{Title}
\usage{
r2hpcc.UnprotectDFUWorkunits(conn, workunits)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{workunits}{- list of workunits to set to unprotected}
}
\value{
- status of processed operation
}
\description{
Title
}
