% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{r2hpcc.Connect}
\alias{r2hpcc.Connect}
\title{Creates a connection object for HPCC}
\usage{
r2hpcc.Connect(host, targetCluster, userId, password, resultLimit,
  debugMode = FALSE)
}
\arguments{
\item{host}{- ip address of HPCC host}

\item{password}{- password associated with the user}

\item{resultLimit}{- limits the result at connection level}

\item{targetcluster}{- Target Thor cluster}

\item{userid}{- user id - if the cluster is secure}
}
\description{
Takes in the HPCC host name, target cluster, user name, password, reult limit
and returns a connection object
}
