% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteDropZoneFile.R
\name{r2hpcc.DeleteDropZoneFile}
\alias{r2hpcc.DeleteDropZoneFile}
\title{This function deletes a file in specified drop zone}
\usage{
r2hpcc.DeleteDropZoneFile(conn, dropZoneFileName, dropZonePath)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{dropZoneFileName}{- file to be deleted}

\item{dropZonePath}{- drop zone path}
}
\value{
- returns result of file uploading
}
\description{
This function deletes a file in specified drop zone
}
