% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFUArrayAction.R
\name{r2hpcc.DeleteLogicalFiles}
\alias{r2hpcc.DeleteLogicalFiles}
\title{This function deletes a list of logical files}
\usage{
r2hpcc.DeleteLogicalFiles(conn, logicalFiles)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{logicalFiles}{- list of logical files}
}
\value{
- status of processed operation
}
\description{
This function deletes a list of logical files
}
