% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFUArrayAction.R
\name{r2hpcc.DFUArrayAction}
\alias{r2hpcc.DFUArrayAction}
\title{This function performs an action on a list of logical files}
\usage{
r2hpcc.DFUArrayAction(conn, action, logicalFiles)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{action}{- action}

\item{logicalFiles}{- list of logical files}
}
\value{
- status of processed operation
}
\description{
This function performs an action on a list of logical files
}
