% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echo.R
\name{r2hpcc.Echo}
\alias{r2hpcc.Echo}
\title{This function takes an input string and "echoes" the value in its result.
This function is intended for end-to-end connectivity testing.
A successful response indicates a good connection to the server hosting the Ws-SQL Web service.}
\usage{
r2hpcc.Echo(conn, echo)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{request}{- String to echo in result}
}
\value{
- Echo string passed in
}
\description{
This function takes an input string and "echoes" the value in its result.
This function is intended for end-to-end connectivity testing.
A successful response indicates a good connection to the server hosting the Ws-SQL Web service.
}
