% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDFUWorkunit.R
\name{r2hpcc.GetDFUWorkunit}
\alias{r2hpcc.GetDFUWorkunit}
\title{This function returns DFU workunit information}
\usage{
r2hpcc.GetDFUWorkunit(conn, workunit)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{workunit}{- DFU workunit id}
}
\value{
- returns workunit
}
\description{
This function returns DFU workunit information
}
