% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWorkunit.R
\name{r2hpcc.GetDFUWorkunit}
\alias{r2hpcc.GetDFUWorkunit}
\title{This function uploads file to specified table in the loading zone}
\usage{
r2hpcc.GetDFUWorkunit(conn, workunit)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{tableName}{- table name where file will be stored. If table exists - it will be overwrtitten}

\item{fileNamePath}{- full path to file to be loaded}

\item{loadingZonePath}{- loading zone path}
}
\value{
- returns result of file uploading
}
\description{
This function uploads file to specified table in the loading zone
}
