% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDropZoneFiles.R
\name{r2hpcc.GetDropZoneFiles}
\alias{r2hpcc.GetDropZoneFiles}
\title{This function returns a list of files in the specified drop zone}
\usage{
r2hpcc.GetDropZoneFiles(conn, dropZoneIP, dropZonePath)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{dropZoneIP}{- drop zone IP}

\item{dropZonePath}{- drop zone path}
}
\value{
- returns a list of files in the drop zone
}
\description{
This function returns a list of files in the specified drop zone
}
