% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDropZoneFiles.R
\name{r2hpcc.GetDropZoneFiles}
\alias{r2hpcc.GetDropZoneFiles}
\title{This function returns a list of drop zone files}
\usage{
r2hpcc.GetDropZoneFiles(conn, loadingZonePath)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{loadingZonePath}{- loading zone path}
}
\value{
- returns result of file uploading
}
\description{
This function returns a list of drop zone files
}
