% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareSQL.R
\name{r2hpcc.PrepareSQL}
\alias{r2hpcc.PrepareSQL}
\title{Calls the executesql function of ws_sql}
\usage{
r2hpcc.PrepareSQL(conn, sqlQuery, timeOut = -1)
}
\arguments{
\item{conn}{- hpcc connection information}

\item{sqlQuery}{- echo message}

\item{timeOut}{- Timeout value in milliseconds. Use -1 for no timeout}
}
\description{
Takes in two parameters HPCC SQL statement and HPCC connection object
Returns a data frame.
If the function eencountered any error while executing the query, 2nd element will have a value of -1 and 1st element, the error message.
}
