% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spray.R
\name{r2hpcc.SprayFixed}
\alias{r2hpcc.SprayFixed}
\title{This function deletes a list of logical files}
\usage{
r2hpcc.SprayFixed(conn, sourceIP, sourcePath, recordSize, destinationGroup,
  namePrefix, targetName, overwrite = FALSE, compress = FALSE)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{sourceIP}{}

\item{sourcePath}{}

\item{recordSize}{}

\item{destinationGroup}{}

\item{namePrefix}{}

\item{targetName}{}

\item{overwrite}{}

\item{compress}{}
}
\value{
- status of processed operation
}
\description{
This function deletes a list of logical files
}
