% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadFile.R
\name{r2hpcc.UploadFile}
\alias{r2hpcc.UploadFile}
\title{This function uploads file to specified table in the drop zone}
\usage{
r2hpcc.UploadFile(conn, destinationFileName, dropZonePath, sourceFileNamePath)
}
\arguments{
\item{conn}{- HPCC connection information}

\item{destinationFileName}{- file name to be used in the drop zone}

\item{dropZonePath}{- drop zone path}

\item{sourceFileNamePath}{- full path to source file}
}
\value{
- result of file uploading
}
\description{
This function uploads file to specified table in the drop zone
}
